fix : Version 코드를 Long 형태로 저장하도록 수정

+++ data/src/main/java/com/yapp/data/model/VersionEntity.kt
@@ -7,9 +7,9 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class VersionEntity(
     @SerialName("min_version_code")
-    val minVersionCode: Int? = null,
+    val minVersionCode: Long? = null,
     @SerialName("max_version_code")
-    val maxVersionCode: Int? = null,
+    val maxVersionCode: Long? = null,
     @SerialName("current_version")
     val currentVersion: String? = null,
 )
