+++ domain/src/main/java/com/yapp/domain/model/Version.kt
@@ -1,7 +1,7 @@
 package com.yapp.domain.model
 
 data class Version(
-    val minVersionCode: Int,
-    val maxVersionCode: Int,
+    val minVersionCode: Long,
+    val maxVersionCode: Long,
     val currentVersion: String,
 )
